code review etienne L
